ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.1",
    "docx2txt>=0.9",
    "fastapi>=0.118.0",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.8",
    "langsmith>=0.4.32",
    "pinecone>=7.3.0",
    "pymongo[srv]==3.12",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "transformers>=4.57.0",
    "uvicorn>=0.37.0",
]
